requirements.txt
# Core API
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6

# RAG & LLM
langchain==0.2.0
langchain-core==0.2.43
langchain-community==0.2.0
langchain-openai==0.1.25
sentence-transformers==3.0.0

# VectorDB
chromadb>=0.5.0
pinecone-client==4.0.0

# MLOps & Tracking
mlflow==2.11.3
langsmith==0.1.147
evidently==0.4.30

# Monitoring
prometheus-client==0.17.1

# Orchestration
apache-airflow==2.9.1

# Evaluation
ragas==0.1.8

# Frontend
streamlit==1.25.0

# Utilities & Data
pymupdf>=1.22.5
python-dotenv==1.0.1
pyyaml==6.0.1
pandas==2.2.2
numpy==1.26.4
requests>=2.32.0

# Testing
pytest==8.2.0
pytest-cov==4.1.0
requests-mock==1.11.0
